feat: Add Fastboot lib, partition flash and FS detection
Introduce a new LibFastbootSharp project with a Fastboot client (connect, command, upload, device enumeration) and add it to the solution and desktop project references. Implement filesystem detection in LibLpSharp.Utility (FilesystemInfo) with support for SquashFS, EROFS, EXT4, F2FS and basic FAT detection, and wire detection into LpDump output. Add PartitionReadStream to stream individual LP partitions (supports linear and zero extents) and a FlashPartitionMessage to send partition streams to the flasher. Update PartitionManager, MainViewModel and FlasherViewModel to support flashing selected partitions (including streaming direct partition data), add UI helpers (EqualityConverter), menu active header binding, UI adjustments and safer command state refreshes. Minor fixes: csproj changes (AllowUnsafeBlocks, package/project refs), small UI/formatting tweaks and a test assertion type correction.

+++ Tests/PartitionToolSharp.Tests/BuilderTests.cs
@@ -102,7 +102,7 @@ public class BuilderTests
         var sparseFile = superBuilder.Build();
 
         Assert.NotNull(sparseFile);
-        Assert.Equal(DeviceSize / 4096, (ulong)sparseFile.Header.TotalBlocks);
+        Assert.Equal(DeviceSize / 4096, sparseFile.Header.TotalBlocks);
         Assert.Contains(sparseFile.Chunks, c => c.Header.ChunkType == SparseFormat.CHUNK_TYPE_RAW);
     }
 }
